+++ Gruntfile.js
@@ -135,6 +135,7 @@ module.exports = function( grunt ) {
                 files: {
                     'public/js/enketo-webform-dev-bundle.js': [ 'public/js/src/main-webform.js' ],
                     'public/js/enketo-webform-edit-dev-bundle.js': [ 'public/js/src/main-webform-edit.js' ],
+                    'public/js/enketo-webform-view-dev-bundle.js': [ 'public/js/src/main-webform-view.js' ],
                     'public/js/enketo-offline-fallback-dev-bundle.js': [ 'public/js/src/main-offline-fallback.js' ]
                 },
                 options: {
@@ -147,6 +148,7 @@ module.exports = function( grunt ) {
                 files: {
                     'public/js/enketo-webform-bundle.js': [ 'public/js/src/main-webform.js' ],
                     'public/js/enketo-webform-edit-bundle.js': [ 'public/js/src/main-webform-edit.js' ],
+                    'public/js/enketo-webform-view-bundle.js': [ 'public/js/src/main-webform-view.js' ],
                     'public/js/enketo-offline-fallback-bundle.js': [ 'public/js/src/main-offline-fallback.js' ]
                 },
             },
@@ -165,6 +167,7 @@ module.exports = function( grunt ) {
                 files: {
                     'public/js/enketo-webform-bundle.min.js': [ 'public/js/enketo-webform-bundle.js' ],
                     'public/js/enketo-webform-edit-bundle.min.js': [ 'public/js/enketo-webform-edit-bundle.js' ],
+                    'public/js/enketo-webform-view-bundle.min.js': [ 'public/js/enketo-webform-view-bundle.js' ],
                     'public/js/enketo-offline-fallback-bundle.min.js': [ 'public/js/enketo-offline-fallback-bundle.js' ],
                 },
             },