+++ app/controllers/submission-controller.js
@@ -17,7 +17,8 @@ module.exports = function( app ) {
 };
 
 router.param( 'enketo_id', routerUtils.enketoId );
-router.param( 'encrypted_enketo_id', routerUtils.encryptedEnketoId );
+router.param( 'encrypted_enketo_id_single', routerUtils.encryptedEnketoIdSingle );
+router.param( 'encrypted_enketo_id_view', routerUtils.encryptedEnketoIdView );
 
 router
     .all( '*', function( req, res, next ) {
@@ -25,10 +26,11 @@ router
         next();
     } )
     .get( '/max-size/:enketo_id', maxSize )
-    .get( '/max-size/:encrypted_enketo_id', maxSize )
+    .get( '/max-size/:encrypted_enketo_id_single', maxSize )
     .get( '/:enketo_id', getInstance )
+    .get( '/:encrypted_enketo_id_view', getInstance )
     .post( '/:enketo_id', submit )
-    .post( '/:encrypted_enketo_id', submit )
+    .post( '/:encrypted_enketo_id_single', submit )
     .all( '/*', function( req, res, next ) {
         var error = new Error( 'Not allowed' );
         error.status = 405;