+++ app/lib/router-utils.js
@@ -1,6 +1,11 @@
 'use strict';
 
 var utils = require( './utils' );
+var config = require( '../models/config-model' ).server;
+var keys = {
+    singleOnce: config[ 'less secure encryption key' ],
+    view: config[ 'less secure encryption key' ] + 'view',
+};
 
 function enketoIdParam( req, res, next, id ) {
     if ( /^::[A-z0-9]{4,8}$/.test( id ) ) {
@@ -11,12 +16,23 @@ function enketoIdParam( req, res, next, id ) {
     }
 }
 
-function encryptedEnketoIdParam( req, res, next, id ) {
+function encryptedEnketoIdParamSingle( req, res, next, id ) {
+    _encryptedEnketoIdParam( req, res, next, id, keys.singleOnce );
+}
+
+function encryptedEnketoIdParamView( req, res, next, id ) {
+    _encryptedEnketoIdParam( req, res, next, id, keys.view );
+}
+
+function _encryptedEnketoIdParam( req, res, next, id, key ) {
     // either 32 or 64 hexadecimal characters
     if ( /^::([0-9a-fA-F]{32}$|[0-9a-fA-F]{64})$/.test( id ) ) {
         req.encryptedEnketoId = id.substring( 2 );
         try {
-            req.enketoId = utils.insecureAes192Decrypt( id.substring( 2 ), req.app.get( 'less secure encryption key' ) );
+            // Just see if it can be decrypted. Storing the encrypted value might
+            // increases chance of leaking underlying enketo_id but for now this is used
+            // in the submission controller and transformation controller.
+            req.enketoId = utils.insecureAes192Decrypt( id.substring( 2 ), key );
             next();
         } catch ( e ) {
             console.error( 'Could not decrypt:', req.encryptedEnketoId );
@@ -27,7 +43,11 @@ function encryptedEnketoIdParam( req, res, next, id ) {
     }
 }
 
+
+
 module.exports = {
     enketoId: enketoIdParam,
-    encryptedEnketoId: encryptedEnketoIdParam
+    idEncryptionKeys: keys,
+    encryptedEnketoIdSingle: encryptedEnketoIdParamSingle,
+    encryptedEnketoIdView: encryptedEnketoIdParamView
 };