+++ public/js/src/module/controller-webform.js
@@ -19,7 +19,7 @@ var formData;
 var $formprogress;
 var formOptions = {
     clearIrrelevantImmediately: false,
-    goTo: settings.type === 'edit' || settings.type === 'preview'
+    goTo: settings.type === 'edit' || settings.type === 'preview' || settings.type === 'view'
 };
 
 function init( selector, data ) {