+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ---------------------
 ##### Added
 - "no-text-transform" style to be used with Grid Theme to prevent uppercasing labels.
+- Readonly views for empty forms and submitted records. See new [/survey/view](https://apidocs.enketo.org/v2#/post-survey-view) and [/instance/view](https://apidocs.enketo.org/v2#/post-instance-view) API endpoints.
 
 [1.47.3] - 2017-06-20
 ---------------------